(MODULES-4153) Fix duplicate resources; unfix umask
staging::file will download a file as the puppet daemon user (root) and
then staging::extract will extract the file as the specified user (eg
tomcat). In the case of strict umasks 0077 the downloaded file would be
root:root 600 and thus inaccessible to the tomcat user. The fix was to
manage the mode from the tomcat module to work around the fact that the
staging module was no longer receiving features. This only worked for
http downloads and caused duplicate declarations for puppet://
downloads.

This patch undoes the umask "fix" since it breaks other download modes.
The appropriate fix at this point is to migrate tomcat to puppet-archive
which does not have this issue, but is backwards-incompatible.